+++ app/controllers/user.js
@@ -20,7 +20,7 @@ exports.create = (req, res, next) => {
     if(error.errors){
       errorBag = error.errors.map(err => err.message);
       logger.error(`A database error occured when attempting a user signup. Details: ${errorBag}.`);
-      res.status(400).send(errorBag);
+      res.status(401).send(errorBag);
     }else{
       logger.error(`Unhandled error! details: ${error}`);
       res.status(500).send(error);
